// simple-rpg-spring-docker/simple-rpg-spring-docker/simple-rpg-ng/src/app/character/model/character-form.ts
/**
 * Represents single edition in form.
 */
export interface CharacterForm {

  /**
   * Name of the character.
   */
  name: string;

  /**
   * Character's background story.
   */
  background: string;

  /**
   * Character's age.
   */
  age: number;

  /**
   * Character's strength.
   */
  strength: number;

  /**
   * Character's constitution.
   */
  constitution: number;

  /**
   * Character's charisma.
   */
  charisma: number;

  /**
   * Character's profession.
   */
  profession: string;

}